requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
pathlib==1.0.1
pillow==10.3.0
xlrd==2.0.1
